backend/requirements.txt
celery==5.2.0
Django==3.2.6
redis==3.5.3
psycopg2-binary==2.8.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
requests==2.32.3